requirements.txt
line-bot-sdk
fastapi
uvicorn[standard]
mysql-connector-python==8.3.0
cryptography # for encrypt sensitive